Update NFT to latest Raphtory model (#598)
* nft update to latest model

* refractor scala example to latest raphtory
update lotr to use latest core
change nft folder to match examples
update python nft example

* add support for jsonformat with comp obj in python

* add wrapper for old graph builder in python

* almost done, just need to fix final type error

* add print statements for debug
fixed bug with incorrect var
fixed str-str comparson to float-float
partial remove dataclass due to pemja issue

* fixed issue with dataclass
fixed issue with -1 not being recognised by pemja
moved tabularise out of dataclass
add filter for nonempty rows

* simplify python code, remove prints

* move out date price map to improve loading times

* nft analysis mostly cleaned, just need to finish brothers and sisters, add more text, then mardown ify with scala version

* python notebook complete with guide

* fix time issue

* remove excess vertex creations
add remote data sources in python
update curl to allow redirection
add curl from osf in scala
use nft_id instead of vertex id

* completed jupyter notebook

* move out pulsar sink for the time being

* move nft code to correct folder
update README.md

* add strange scaladoc reader causing classpath errors

* remove bad comment in md

* spelling mistake

* added readthedoc example up to the analysis part.

* completed tutorial
