+++ examples/nft/build.sbt
@@ -2,7 +2,8 @@ name := "example-nft"
 version := "0.5"
 organization := "com.raphtory"
 scalaVersion := "2.13.7"
-libraryDependencies += "com.raphtory" %% "core" % "0.1.0"
+libraryDependencies += "com.raphtory" %% "core" % "0.2.0a1"
 resolvers += Resolver.mavenLocal
 Compile / resourceDirectory := baseDirectory.value / "resources"
 libraryDependencies += "net.sf.py4j" % "py4j" % "0.10.9.5"
+libraryDependencies += "com.github.takezoe" %% "runtime-scaladoc-reader" % "1.0.3"
