+++ examples/nft/project/build.properties
