+++ python/pyraphtory/pyraphtory/formats.py
@@ -1,0 +1,5 @@
+from pyraphtory.interop import ScalaClassProxy
+
+
+class JsonFormat(ScalaClassProxy):
+    _classname = "com.raphtory.formats.JsonFormat"
